+++ examples/resources/msgraph_resource/resource.tf
@@ -28,3 +28,8 @@ output "all" {
   // it will output the whole response
   value = msgraph_resource.application.output.all
 }
+
+output "resource_url" {
+  // it will output something like "applications/12345678-1234-1234-1234-123456789abc"
+  value = msgraph_resource.application.resource_url
+}